Refactor to old way of doing things
