+++ src/Services/WSD.Catalog.Domain/Logic/CatalogItemLogic.cs
@@ -1,0 +1,49 @@
+using WSD.Catalog.Domain.Exceptions;
+using WSD.Catalog.Domain.Models;
+
+namespace WSD.Catalog.Domain.Logic;
+
+public class CatalogItemLogic : CatalogItem, ICatalogItemLogic
+{
+    /// <inheritdoc />
+    public int RemoveStock(int quantityDesired)
+    {
+        if (AvailableStock == 0)
+        {
+            throw new CatalogDomainException($"Empty stock, product item {Name} is sold out");
+        }
+
+        if (quantityDesired <= 0)
+        {
+            throw new CatalogDomainException($"Item units desired should be greater than zero");
+        }
+
+        int removed = Math.Min(quantityDesired, AvailableStock);
+
+        AvailableStock -= removed;
+
+        return removed;
+    }
+
+    /// <inheritdoc />
+    public int AddStock(int quantity)
+    {
+        int original = AvailableStock;
+
+        // The quantity that the client is trying to add to stock is greater than what can be physically accommodated in the Warehouse
+        if (AvailableStock + quantity > MaxStockThreshold)
+        {
+            // For now, this method only adds new units up maximum stock threshold. In an expanded version of this application, we
+            //could include tracking for the remaining units and store information about overstock elsewhere. 
+            AvailableStock += MaxStockThreshold - AvailableStock;
+        }
+        else
+        {
+            AvailableStock += quantity;
+        }
+
+        OnReorder = false;
+
+        return AvailableStock - original;
+    }
+}
