+++ src/Services/WSD.Catalog.Domain/Models/CatalogType.cs
@@ -2,8 +2,6 @@
 {
     public class CatalogType
     {
-        public int Id { get; set; }
-
         public string Type { get; set; }
     }
 }
