+++ src/Services/WSD.Catalog.Infrastructure/Models/IBaseEntity.cs
@@ -1,0 +1,9 @@
+namespace WSD.Catalog.Infrastructure.Models;
+
+public interface IBaseEntity
+{
+    /// <summary>
+    /// Primary key
+    /// </summary>
+    public int Id { get; set; }
+}
