+++ src/Services/WSD.Catalog.Infrastructure/WSD.Catalog.Infrastructure.csproj
@@ -12,4 +12,8 @@
     <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.11" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\WSD.Catalog.Domain\WSD.Catalog.Domain.csproj" />
+  </ItemGroup>
+
 </Project>
